requirements.txt
# Core Flask and Web Server
flask
flask-wtf
flask-limiter
gunicorn
werkzeug

# Database
psycopg2-binary

# Utilities
python-dotenv
requests

# Payments
stripe>=8.0.0

# AI Provider SDKs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
together>=1.0.0

# Testing
pytest>=7.0.0
